requirements.txt
gradio==5.16.0
openai==1.62.0
pandas==2.2.3
python-dotenv==1.0.1
